Used standard file structs
